Rename ons cleaned column names file
