+++ utils/column_values/categorical_columns_by_dataset.py
@@ -13,7 +13,7 @@ from utils.column_names.cleaned_data_files.cqc_provider_cleaned import (
 from utils.column_names.cleaned_data_files.ascwds_worker_cleaned import (
     AscwdsWorkerCleanedColumns as AWKClean,
 )
-from utils.column_names.cleaned_data_files.ons_cleaned_values import (
+from utils.column_names.cleaned_data_files.ons_cleaned import (
     OnsCleanedColumns as ONSClean,
     ONScol as ONS,
 )
